+++ EssentialFeed/EssentialFeediOS/Feed ImageLoader/FeedImageDataLoader.swift
@@ -1,0 +1,18 @@
+//
+//  FeedImageDataLoader.swift
+//  EssentialFeediOS
+//
+//  Created by AmritPandey on 19/07/22.
+//
+
+import Foundation
+
+
+public protocol FeedImageDataLoaderTask {
+    func cancel()
+}
+
+public protocol FeedImageDataLoader {
+    typealias Result = Swift.Result<Data, Error>
+    func loadImageData(from url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoaderTask
+}
